03-career-conversations/requirements.txt
python-dotenv
openai
pypdf
gradio
requests